fix shared types
