+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@gideonparanoid/solax",
+  "name": "solax",
   "version": "1.0.0",
   "private": true,
   "scripts": {
