+++ packages/common/package.json
@@ -1,7 +1,8 @@
 {
-  "name": "solax-poller",
+  "name": "solax-common",
   "version": "1.0.0",
   "private": true,
+  "main": "dist",
   "scripts": {
     "lint": "eslint src/**/*.{js,ts}",
     "lint:fix": "yarn lint --fix",
@@ -10,22 +11,15 @@
   },
   "license": "ISC",
   "dependencies": {
-    "agenda": "^3.1.0",
-    "axios": "^0.19.2",
-    "mongodb": "^3.5.7",
-    "solax-common": "1.0.0"
+    "dotenv": "^8.2.0"
   },
   "devDependencies": {
-    "@types/agenda": "^2.0.7",
-    "@types/express": "^4.17.6",
-    "@types/mongodb": "^3.5.17",
     "@types/node": "^14.0.1",
     "@typescript-eslint/eslint-plugin": "^2.33.0",
     "@typescript-eslint/parser": "^2.33.0",
     "eslint": "^7.0.0",
     "eslint-config-airbnb-base": "^14.1.0",
     "eslint-config-prettier": "^6.11.0",
-    "eslint-import-resolver-typescript": "^2.0.0",
     "eslint-plugin-import": "^2.20.2",
     "eslint-plugin-prettier": "^3.1.3",
     "prettier": "^2.0.5",
