+++ packages/common/src/config.ts
@@ -1,0 +1,12 @@
+import * as dotenv from 'dotenv';
+
+const parsed = dotenv.config().parsed;
+
+const solaxURL = parsed.solax_url;
+
+const db = {
+  url: parsed.db_url,
+  name: 'solax',
+};
+
+export { solaxURL, db };
