+++ packages/poller/src/db.ts
@@ -1,6 +1,6 @@
 import { Db } from 'mongodb';
 
-import { UsefulRTData, UsefulDailyData } from '@types';
+import { UsefulRTData, UsefulDailyData } from 'solax-common/types';
 
 export class SolaxDb {
   private db: Db;
