+++ packages/poller/tsconfig.json
@@ -12,7 +12,7 @@
       "*": [
         "node_modules/*"
       ],
-      "@types": [ "../types/" ]
+      "solax-common*": [ "../common/src*" ]
     }
   },
   "include": [
