+++ packages/server/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@gideonparanoid/solax-server",
+  "name": "solax-server",
   "version": "1.0.0",
   "private": true,
   "scripts": {
@@ -22,6 +22,7 @@
     "eslint": "^7.0.0",
     "eslint-config-airbnb-base": "^14.1.0",
     "eslint-config-prettier": "^6.11.0",
+    "eslint-import-resolver-typescript": "^2.0.0",
     "eslint-plugin-import": "^2.20.2",
     "eslint-plugin-prettier": "^3.1.3",
     "prettier": "^2.0.5",
